Linters errors

+++ src/app/api/menu-items/route.js
@@ -1,32 +1,41 @@
 import { MenuItems } from "../../models/MenuItems";
 import mongoose from "mongoose";
+import { isAdmin } from "../auth/[...nextauth]/route";
 
-export async function POST (req) {
-    mongoose.connect(process.env.MONGO_URL)
-    const data = await req.json();
+export async function POST(req) {
+  mongoose.connect(process.env.MONGO_URL);
+  const data = await req.json();
+  if (await isAdmin()) {
     const menuItemDoc = await MenuItems.create(data);
     return Response.json(menuItemDoc);
+  } else {
+    return Response.json({});
+  }
 }
 
-export async function PUT (req) {
-    mongoose.connect(process.env.MONGO_URL)
-    const {_id, ...data} = await req.json();
+export async function PUT(req) {
+  mongoose.connect(process.env.MONGO_URL);
+  if (await isAdmin()) {
+    const { _id, ...data } = await req.json();
     await MenuItems.findByIdAndUpdate(_id, data);
-    return Response.json(true);
+  }
+  return Response.json(true);
 }
 
 export async function GET() {
-    mongoose.connect(process.env.MONGO_URL);
-    return Response.json(
-        await MenuItems.find()
-    );
-
+  mongoose.connect(process.env.MONGO_URL);
+    return Response.json(await MenuItems.find());
+    return Response.json([]);
 }
 
 export async function DELETE(req) {
-    mongoose.connect(process.env.MONGO_URL);
-    const url = new URL(req.url);
-    const _id = url.searchParams.get('_id');
-    await MenuItems.deleteOne({_id});
-    return Response.json(true);
+  mongoose.connect(process.env.MONGO_URL);
+  const url = new URL(req.url);
+  const _id = url.searchParams.get("_id");
+
+  if (await isAdmin()) {
+    await MenuItems.deleteOne({ _id });
+  }
+
+  return Response.json(true);
 }
